Add PKCS#8 PEM encoded private key
This also renames the files and env-variables to make them follow a consistent pattern. Unfortunately, this would require updating all secrets, both certificate and user, which doesn't normally happen in sync and usually only on creation of the database.

For that reason, this can't be done directly as implemented here, and would also need some kind of conversion of the existing secrets coordinated with the rollout of this change.